etl presence-absence: move funcs to shared module
Move the `upsert_presence_absence` and `find_or_create_target` functions
from the `etl presence-absence` command into a shared module for reuse
by future etl commands.
